planes_recogniser/requirements.txt
cycler==0.11.0
kiwisolver==1.4.4
matplotlib==3.4.2
numpy==1.21.0
Pillow==8.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
torch==2.0.0
torchvision==0.15.1
typing-extensions==4.5.0
PyQt5==5.14.0

albumentations
pandas
tqdm
